refrac: made sections its own component to clean up index file
